Allow fields to control their own row
