+++ components/General/SchemaFormFields/Text.js
@@ -2,6 +2,8 @@ import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 import { TextInput, StyleSheet } from 'react-native';
 
+import FormRow from '../FormRow';
+
 const styles = StyleSheet.create({
 	container: {
 		flex: 1,
@@ -25,13 +27,15 @@ export default class Text extends Component {
 			? this.props.schema.default
 			: placeholder;
 		return (
-			<TextInput
-				value={this.props.value}
-				style={styles.container}
-				placeholder={placeholder}
-				onChangeText={this.props.onChange}
-				onSubmitEditing={this.props.onSave}
-			/>
+			<FormRow label={ this.props.name }>
+				<TextInput
+					value={this.props.value}
+					style={styles.container}
+					placeholder={placeholder}
+					onChangeText={this.props.onChange}
+					onSubmitEditing={this.props.onSave}
+				/>
+			</FormRow>
 		);
 	}
 }
